최근 러닝데이터 3개 반환하는 viewRecentRunning api 개발 완료

See merge request swm-13-main/13_swm21/sprint-be!17

+++ src/main/java/sprint/server/domain/statistics/Statistics.java
@@ -24,7 +24,7 @@ public class Statistics {
     private StatisticsType statisticsType;
 
     private double distance;
-    private int totalSeconds;
+    private double totalSeconds;
     private int count;
     private double energy;
